edu/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-jazzmin==2.6.0
django-payments==1.0.0
django-payments-razorpay==1.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
geographiclib==2.0
idna==3.4
phonenumberslite==8.13.4
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.7
razorpay==1.3.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
